Email-Processor-WorkFlow-RAG-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
scikit-learn==1.5.2
numpy==2.1.2
pydantic==2.9.2
google-generativeai==0.8.3


